+++ chain/index.js
@@ -1,0 +1,4 @@
+const getNetwork = require('./get_network');
+const getTransactionRecord = require('./get_transaction_record');
+
+module.exports = {getNetwork, getTransactionRecord};