+++ graph/index.js
@@ -1,5 +1,6 @@
 const getGraphNode = require('./get_graph_node');
 const getGraphPair = require('./get_graph_pair');
 const getNodeAlias = require('./get_node_alias');
+const getScoredNodes = require('./get_scored_nodes');
 
-module.exports = {getGraphNode, getGraphPair, getNodeAlias};
+module.exports = {getGraphNode, getGraphPair, getNodeAlias, getScoredNodes};