+++ peers/index.js
@@ -1,4 +1,5 @@
 const findKey = require('./find_key');
+const getLiquidity = require('./get_liquidity');
 const getPeerLiquidity =  require('./get_peer_liquidity');
 
-module.exports = {findKey, getPeerLiquidity};
+module.exports = {findKey, getLiquidity, getPeerLiquidity};